feat(Component): Convert functionality into a reusable component
