+++ src/renderer/components/main/history.tsx
@@ -1,0 +1,7 @@
+export default function History() {
+  return (
+    <h1 id="testText" className="text-slate-950">
+      검사 기록입니다.
+    </h1>
+  );
+}
