+++ src/renderer/components/test/testForm.tsx
@@ -1,8 +1,12 @@
-import ico_refresh from '@assets/images/icons/icon_refresh.png';
-import ico_check from '@assets/images/icons/icon_check.png';
+import { useState } from 'react';
+
 import ico_angle_left from '@assets/images/icons/icon_angle_left.png';
 import ico_angle_right from '@assets/images/icons/icon_angle_right.png';
-import { useState } from 'react';
+
+import ExamineeInfoForm from './examineeInfoForm';
+import PreCheckScreen from './preCheckScreen';
+import CheckScreen from './checkScreen';
+import TestResult from './testResult';
 
 export default function TestForm() {
   let [currentPage, setCurrentPage] = useState(0);
@@ -61,436 +65,5 @@ export default function TestForm() {
       {currentPage === 2 && <CheckScreen />}
       {currentPage === 3 && <TestResult />}
     </div>
-  );
-}
-
-// 검사 시작 화면
-function ExamineeInfoForm() {
-  return (
-    <>
-      <h1 className="text-slate-950 text-center mb-10">
-        피검사자 기본 정보를 입력해 주세요.
-      </h1>
-
-      <div className="input-wrapper text-slate-950">
-        <ul className="flex flex-col">
-          <li className="flex justify-between items-center mb-5">
-            <p className="w-48">이름</p>
-            <input
-              type="text"
-              name="username"
-              id="username"
-              className="border-b border-gray-300 focus:border-blue-500 w-full"
-            />
-          </li>
-          <li className="flex justify-between items-center mb-5">
-            <p className="w-48">성별</p>
-            <input
-              type="text"
-              name="gender"
-              id="gender"
-              className="border-b border-gray-300 focus:border-blue-500 w-full"
-            />
-          </li>
-          <li className="flex justify-between items-center mb-5">
-            <p className="w-48">생년월일</p>
-            <input
-              type="text"
-              name="birthDate"
-              id="birthDate"
-              className="border-b border-gray-300 focus:border-blue-500 w-full"
-            />
-          </li>
-          <li className="flex justify-between items-center mb-5">
-            <p className="w-48">피검사자번호</p>
-            <input
-              type="text"
-              name="examineeId"
-              id="examineeId"
-              className="border-b border-gray-300 focus:border-blue-500 w-full"
-            />
-          </li>
-          <li className="flex justify-between items-center mb-5">
-            <label htmlFor="direction-select" className="w-48">
-              제시방향
-            </label>
-            <select id="direction-select" className="w-full">
-              <option value="좌">좌</option>
-              <option value="우">우</option>
-              <option value="양방향">양방향</option>
-              <option value="좌 노이즈 우 스피치">좌 노이즈 우 스피치</option>
-              <option value="우 노이즈 좌 스피치">우 노이즈 좌 스피치</option>
-            </select>
-          </li>
-          <li className="flex justify-between items-center mb-5">
-            <label htmlFor="start-level" className="w-48">
-              시작레벨
-            </label>
-            <select id="start-level" className="w-full">
-              <option value="-6">-6</option>
-              <option value="-5">-5</option>
-              <option value="-4">-4</option>
-              <option value="-3">-3</option>
-              <option value="-2">-2</option>
-              <option value="-1">-1</option>
-              <option value="0">0</option>
-              <option value="+1">+1</option>
-              <option value="+2">+2</option>
-              <option value="+3">+3</option>
-              <option value="+4">+4</option>
-              <option value="+5">+5</option>
-              <option value="+6">+6</option>
-              <option value="+7">+7</option>
-              <option value="+8">+8</option>
-              <option value="+9">+9</option>
-              <option value="+10">+10</option>
-            </select>
-          </li>
-          <li className="flex justify-between items-center mb-5">
-            <label htmlFor="scoring-method" className="w-48">
-              채점방식
-            </label>
-            <select id="scoring-method" className="w-full">
-              <option value="Digit Scoring">Digit Scoring</option>
-              <option value="Tripet Scoring">Tripet Scoring</option>
-            </select>
-          </li>
-          <li className="flex justify-between items-center mb-5">
-            <label htmlFor="sound-set" className="w-48">
-              사운드세트
-            </label>
-            <select id="sound-set" className="w-full">
-              <option value="List 1">List 1</option>
-              <option value="List 2">List 2</option>
-              <option value="List 3">List 3</option>
-              <option value="List 4">List 4</option>
-              <option value="List 5">List 5</option>
-              <option value="List 6">List 6</option>
-            </select>
-          </li>
-          <li className="flex flex-col flex-start mb-5">
-            <p className="w-48">참고사항</p>
-            <textarea
-              className="w-full h-32 border border-gray-300 shadow-md p-2 resize-none"
-              placeholder="메모를 입력하세요."
-            />
-          </li>
-          <div className="refresh-check-wrapper flex justify-between items-center">
-            <a
-              href="#!"
-              className="flex justify-center items-center rounded-full bg-gray-300 w-20 h-20 mx-auto"
-            >
-              <img src={ico_refresh} alt="ico_refresh" className="w-1/2" />
-            </a>
-            <a
-              href="#!"
-              className="flex justify-center items-center rounded-full bg-gray-300 w-20 h-20 mx-auto"
-            >
-              <img src={ico_check} alt="ico_check" className="w-1/2" />
-            </a>
-          </div>
-        </ul>
-      </div>
-    </>
-  );
-}
-
-// 사전검사 화면
-function PreCheckScreen() {
-  const [value, setValue] = useState(0);
-
-  const handleSliderChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const newValue = parseInt(e.target.value);
-    setValue(newValue);
-  };
-
-  const [digits, setDigits] = useState(['', '', '']);
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  const handleNumberClick = (number: string) => {
-    if (currentIndex < 3) {
-      const updatedDigits = [...digits];
-      updatedDigits[currentIndex] = number;
-      setDigits(updatedDigits);
-      setCurrentIndex(currentIndex + 1);
-    }
-  };
-
-  const handleConfirmClick = () => {
-    const enteredNumber = digits.join('');
-    alert('입력한 숫자: ' + enteredNumber);
-  };
-
-  const renderButtons = () => {
-    const buttons = [];
-    for (let i = 1; i <= 9; i++) {
-      buttons.push(
-        <div key={i} className="w-1/3 p-1">
-          <button
-            className="rounded-full bg-gray-300 w-16 h-16 text-2xl"
-            onClick={() => handleNumberClick(i.toString())}
-          >
-            {i}
-          </button>
-        </div>
-      );
-    }
-    // 간격 조절용 버튼
-    buttons.push(
-      <div key={-1} className="w-1/3">
-        <button className="hidden" disabled>
-          공백
-        </button>
-      </div>
-    );
-    // 간격 조절용 버튼 끝
-    buttons.push(
-      <div key={0} className="w-1/3">
-        <button
-          className="rounded-full bg-gray-300 w-16 h-16 text-2xl"
-          onClick={() => handleNumberClick('0')}
-        >
-          0
-        </button>
-      </div>
-    );
-    buttons.push(
-      <div key="confirm" className="w-1/3">
-        <button
-          className="rounded-full bg-gray-300 w-16 h-16 text-2xl"
-          onClick={handleConfirmClick}
-        >
-          &#10003;
-        </button>
-      </div>
-    );
-    return buttons;
-  };
-
-  return (
-    <>
-      <h1 className="text-slate-950 text-center -mt-5 mb-10">
-        연습 검사를 실시합니다. <br />
-        검사 전 아래 레버를 움직여 최적의 소리 강도를 찾아내세요.
-      </h1>
-
-      <div className="input-wrapper text-slate-950 mb-10">
-        <div className="flex flex-col justify-center items-center">
-          <div className="flex w-full">
-            <span className="mr-3">0dB</span>
-            <input
-              type="range"
-              min={0}
-              max={100}
-              step={5}
-              value={value}
-              onChange={handleSliderChange}
-            />
-            <span className="ml-3">100dB</span>
-          </div>
-          <span className="text-slate-950">{value}dB</span>
-        </div>
-      </div>
-
-      <div className="text-center text-slate-950">
-        <p>이제 3개의 연속된 숫자가 들리게 됩니다.</p>
-        <p>숫자를 다 듣고 해당 숫자를 순서대로 말하세요.</p>
-        <p>반드시 3개의 숫자가 다 제시된 후 말하세요.</p>
-        <p>잘 듣지 못한 경우, 추측해서 숫자 3개를 모두 말해야 합니다.</p>
-      </div>
-
-      <div className="flex flex-col justify-center items-center text-center text-slate-950 w-64 mx-auto">
-        <div className="flex mt-3 mb-5">
-          {digits.map((digit, index) => (
-            <input
-              key={index}
-              type="text"
-              value={digit}
-              readOnly
-              className="w-1/3 text-center p-2 text-2xl border-b-2 border-black ml-3"
-            />
-          ))}
-        </div>
-        <div className="flex flex-wrap">{renderButtons()}</div>
-      </div>
-
-      <div className="flex justify-center items-center text-slate-950 mt-5">
-        <button
-          type="button"
-          className="w-36 h-14 bg-green-200 rounded-full mr-10"
-        >
-          시작
-        </button>
-        <button type="button" className="w-36 h-14 bg-gray-300	 rounded-full">
-          완료
-        </button>
-      </div>
-    </>
-  );
-}
-
-function CheckScreen() {
-  const [digits, setDigits] = useState(['', '', '']);
-  const [currentIndex, setCurrentIndex] = useState(0);
-  const [count, setCount] = useState(1);
-  const [error, setError] = useState(false);
-  const totalQuestions = 25;
-
-  const handleNumberClick = (number: string) => {
-    if (currentIndex < 3) {
-      const updatedDigits = [...digits];
-      updatedDigits[currentIndex] = number;
-      setDigits(updatedDigits);
-      setCurrentIndex(currentIndex + 1);
-    }
-  };
-
-  const handleCheck = () => {
-    if (currentIndex === 3) {
-      if (count < totalQuestions) {
-        setCount(count + 1);
-        setDigits(['', '', '']);
-        setCurrentIndex(0);
-        setError(false);
-      }
-    } else {
-      setError(true);
-      alert('모든 숫자가 입력되지 않았습니다.');
-    }
-  };
-
-  const renderButtons = () => {
-    const buttons = [];
-    for (let i = 1; i <= 9; i++) {
-      buttons.push(
-        <div key={i} className="w-1/3 p-1">
-          <button
-            className="rounded-full bg-gray-300 w-16 h-16 text-2xl"
-            onClick={() => handleNumberClick(i.toString())}
-          >
-            {i}
-          </button>
-        </div>
-      );
-    }
-    // 간격 조절용 버튼
-    buttons.push(
-      <div key={-1} className="w-1/3">
-        <button className="hidden" disabled>
-          공백
-        </button>
-      </div>
-    );
-    // 간격 조절용 버튼 끝
-    buttons.push(
-      <div key={0} className="w-1/3">
-        <button
-          className="rounded-full bg-gray-300 w-16 h-16 text-2xl"
-          onClick={() => handleNumberClick('0')}
-        >
-          0
-        </button>
-      </div>
-    );
-    buttons.push(
-      <div key="confirm" className="w-1/3">
-        <button
-          className="rounded-full bg-gray-300 w-16 h-16 text-2xl"
-          onClick={handleCheck}
-        >
-          &#10003;
-        </button>
-      </div>
-    );
-    return buttons;
-  };
-
-  return (
-    <>
-      <h1 className="text-slate-950 text-center mb-11">테스트 진행 중</h1>
-
-      {/* 오른쪽 SNB 시작 */}
-      <div className="flex flex-col absolute end-0 text-slate-950 bg-white border-2 border-black box-border">
-        <div className="bg-green-200 p-5 border-b-2 border-black">
-          <p>검사현황</p>
-        </div>
-        <div className="p-5">
-          <p>검사 사운드 세트: List 1</p>
-          <p className="mt-5 mb-5">
-            사운드 제시 방향: Noise Left + Speech Right
-          </p>
-          <p>채점 방식: Digit Scoring</p>
-        </div>
-      </div>
-      {/* 오른쪽 SNB 끝 */}
-
-      <div className="text-center text-slate-950 mt-16">
-        <p>
-          {count}/{totalQuestions}
-        </p>
-      </div>
-
-      <div className="flex flex-col justify-center items-center text-center text-slate-950 w-64 mx-auto mt-24">
-        <div className="flex mt-3 mb-5">
-          {digits.map((digit, index) => (
-            <input
-              key={index}
-              type="text"
-              value={digit}
-              readOnly
-              className="w-1/3 text-center p-2 text-2xl border-b-2 border-black ml-3"
-            />
-          ))}
-        </div>
-        <div className="flex flex-wrap">{renderButtons()}</div>
-      </div>
-
-      <div className="flex justify-center items-center text-slate-950 mt-5">
-        <button type="button" className="w-36 h-14 bg-green-200 rounded-full">
-          중단하기
-        </button>
-      </div>
-    </>
-  );
-}
-
-function TestResult() {
-  return (
-    <>
-      <h1 className="text-slate-950 text-center mb-11">테스트 결과</h1>
-
-      {/* 오른쪽 SNB 시작 */}
-      <div className="flex flex-col absolute end-0 text-slate-950 bg-white border-2 border-black box-border">
-        <div className="bg-green-200 p-5 border-b-2 border-black">
-          <p>검사현황</p>
-        </div>
-        <div className="p-5">
-          <p>검사 사운드 세트: List 1</p>
-          <p className="mt-5 mb-5">
-            사운드 제시 방향: Noise Left + Speech Right
-          </p>
-          <p>채점 방식: Digit Scoring</p>
-        </div>
-      </div>
-      {/* 오른쪽 SNB 끝 */}
-
-      <div className="graph-wrapper flex justify-center items-center text-slate-950">
-        <div className="flex justify-center items-center border border-black w-1/2 h-96">
-          그래프 영역
-        </div>
-      </div>
-
-      <div className="flex justify-between items-center text-slate-950 mt-10">
-        <button type="button" className="w-48 h-14 bg-green-200 rounded-full">
-          내 결과와 오버레이
-        </button>
-        <button type="button" className="w-48 h-14 bg-green-200 rounded-full">
-          PDF로 저장
-        </button>
-        <button type="button" className="w-48 h-14 bg-green-200 rounded-full">
-          저장 후 종료
-        </button>
-      </div>
-    </>
   );
 }
